service-workers/service-worker/websocket-in-service-worker.https.html

<!DOCTYPE html>
<title>Service Worker: WebSockets can be created in a Service Worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
  const SCRIPT = 'resources/websocket-worker.js?pipe=sub';
  const SCOPE = 'resources/blank.html';

  const registration = await service_worker_unregister_and_register(
    t,
    SCRIPT,
    SCOPE
  );

  t.add_cleanup(async() => {
    if (registration) await registration.unregister();
  });

  await wait_for_state(t, registration.installing, 'activated');
  await new Promise(resolve => {
    navigator.serviceWorker.onmessage = t.step_func(msg => {
      assert_equals(msg.data, 'PASS');
      resolve();
    });
    registration.active.postMessage({});
  });
}, 'Verify WebSockets can be created in a Service Worker');
</script>
